+++ CHANGELOG.md
@@ -3,6 +3,22 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [2.1.0] - 2021-08-16
+### Added
+- Editor Coroutines v1.0.0 package dependency
+- GameViewSizeHelper.cs + MIT License, for changing game view resolution, credit to https://github.com/Syy9/GameViewSizeChanger
+- Editor.asmedf GUID Unity.EditorCoroutines.Editor
+
+### Changed
+- Screenshot.ProcessScreenshot() changed to IEnumerator
+- Screenshots are now placed in a queue and taken when possible (cause of Game View + UI needs to wait a couple frames)
+
+### Fixed
+- Bug where camera wasn't assigned at startup
+- Resolution not being added to file name when including UI
+- Game View + UI resolution not working correctly
+- Generating multiple screenshots now have correct screenshot size assigned
+
 ## [2.0.0] - 2021-08-16
 ### Changed
 - Major overhaul of all scripts and editor window look.